@@ -41,7 +41,7 @@ dependencies:
   path_provider: ^2.0.8
   fl_chart: ^0.67.0
   #background_fetch: ^1.3.3
-  font_awesome_flutter: ^9.1.0
+  font_awesome_flutter: ^10.7.0
   awesome_notifications_core: ^0.9.0
   awesome_notifications: any
 
